refactor: modularize conversion and enhancers
